# docs/api.yaml
openapi: 3.0.0
info:
  version: "1.0.0"
  title: 'Wallet customization service'
security:
  - bearerAuth: []
servers:
  - url: /
  - url: http://127.0.0.1
paths:
...
  /customization/public/v1/logo-settings:
    get:
      security:
        - bearerAuth: []
      tags:
        - Customization
      summary: Shows a logo settings.
      description: The settings include logo path, points for cropping and the fav.icon path.
      operationId: getLogoSettings
      responses:
        '200':
          description: Successful request
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/LogoSettings'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenResponse'
...
components:
  schemas:
...
    LogoSettings:
      type: object
      properties:
        logo:
          type: object
          properties:
            path:
              type: string
              example: "http://example.com/accounts/logo.jpeg"
            xPoint:
              type: integer
              example: 10
            yPoint:
              type: integer
              example: 10
            width:
              type: integer
              example: 200
            height:
              type: integer
              example: 200
        favicon:
          type: object
          properties:
            path:
              type: string
              example: "http://example.com/accounts/fav.icon"
...
    BadRequestResponse:
      type: object
      properties:
        errors:
          type: array
          items:
            $ref: '#/components/schemas/BadRequestError'
    BadRequestError:
      type: object
      properties:
        title:
          type: string
          example: Bad Request
        details:
          type: string
          example: Bad Request
        target:
          type: string
          example: common
    ForbiddenResponse:
      type: object
      properties:
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ForbiddenError'
    ForbiddenError:
      type: object
      properties:
        title:
          type: string
          example: Forbidden
        details:
          type: string
          example: Forbidden
        target:
          type: string
          example: common
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT